flaskServer/requirements.txt
flask
python-dotenv
pyjwt
requests
pytz
logging
statistics
pandas
scikit-learn
joblib